+++ tests/Tests.hs
@@ -10,14 +10,12 @@ import qualified Tests.Lift as Lift
 import qualified Tests.Properties as Properties
 import qualified Tests.Regressions as Regressions
 
-import qualified Tests.Inspection.Strict as InspectionStrict
-import qualified Tests.Inspection.Lazy   as InspectionLazy
+-- import qualified Tests.Inspection.Strict as InspectionStrict
+-- import qualified Tests.Inspection.Lazy   as InspectionLazy
 
 main :: IO ()
 main = defaultMain $ testGroup "All"
   [ Lift.tests
   , Properties.tests
   , Regressions.tests
-  , InspectionStrict.tests
-  , InspectionLazy.tests
   ]